+++ src/lib/pages/block-details/index.tsx
@@ -19,7 +19,8 @@ const BlockDetail = () => {
   useEffect(() => {
     if (router.isReady) AmpTrack(AmpEvent.TO_BLOCK_DETAIL);
   }, [router.isReady]);
-  if (isLoading) return <Loading />;
+
+  if (isLoading) return <Loading withBorder />;
 
   return (
     <PageContainer>