+++ src/lib/services/validatorService.ts
@@ -25,7 +25,11 @@ export const useValidator = (
     getValidator(queryKey[2], queryKey[3] as ValidatorAddr);
 
   return useQuery(
-    ["query", "validator", lcdEndpoint, validatorAddr] as string[],
+    [
+      CELATONE_QUERY_KEYS.VALIDATOR_INFO_BY_ADDRESS,
+      lcdEndpoint,
+      validatorAddr,
+    ] as string[],
     queryFn,
     {
       enabled: enabled && Boolean(validatorAddr),
@@ -74,8 +78,7 @@ export const useValidatorImage = (
 
   return useQuery({
     queryKey: [
-      "query",
-      "validator_identity",
+      CELATONE_QUERY_KEYS.VALIDATOR_IDENTITY_BY_ADDRESS,
       chainName,
       validator?.validatorAddress,
     ],